Refactor the fixtures so they are accessible by new test_consistency
